backend/requirements3.txt
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coverage==7.3.2
cryptography==41.0.5
defusedxml==0.7.1
Django==4.2.7
django-appconf==1.0.6
django-autoslug==1.9.9
django-cors-headers==4.3.1
django-extensions==3.2.3
django-imagekit==5.0.0
django-stubs==4.2.6
django-stubs-ext==4.2.5
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.2
drf-spectacular==0.26.5
exceptiongroup==1.2.0
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
kaitaistruct==0.10
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==23.2
pilkit==3.0
Pillow==10.1.0
pluggy==1.3.0
progressbar2==4.2.0
psycopg2==2.9.9
pyasn1==0.5.1
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
PySocks==1.7.1
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.7.0
python-dotenv==1.0.0
python-utils==3.8.1
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
requests==2.31.0
requests-mock==1.11.0
requests-oauthlib==1.3.1
rpds-py==0.13.1
selenium==4.15.2
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
tomli==2.0.1
trio==0.23.1
trio-websocket==0.11.1
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
typing_extensions==4.8.0
Unidecode==1.3.7
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.24.0.post1
webdriver-manager==4.0.1
wsproto==1.2.0
zstandard==0.22.0
django-debug-toolbar
ipinfo-django
